Fixed a problem
Was having a problem with an exception error where the program wasn't able to find a picture of the track, and the screen kept coming up blank. That's been fixed now, but I've regressed to a previous version of the game controller, so maybe I can look at a previous copy to get a more advanced version of the beatEmUpController.